bot/pyproject.toml
[project]
name = "lms-bot"
version = "0.1.0"
description = "Telegram bot for the LMS backend"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
bot = "bot.bot:main"

[tool.uv]
# Allow running bot.py directly with `uv run bot.py`
